Extract the permission filter into its own class
