+++ lib/actions/index.js
@@ -239,7 +239,6 @@ class ActionRequestWorker {
 
 		debug(`Executing internal action ${actionCard.slug}`)
 		return actionFunction(this.jellyfish, this, targetCard, {
-			actor: await this.jellyfish.getSessionUser(this.session),
 			timestamp: time.getCurrentTimestamp(),
 			session: this.session
 		}, compiledArguments)
