+++ lib/actions/library.js
@@ -82,6 +82,12 @@ module.exports = {
 		})
 	},
 	'action-create-event': async (jellyfish, worker, card, context, options) => {
+		const sessionCard = await jellyfish.getCardById(context.session, context.session)
+
+		if (!sessionCard) {
+			throw new jellyfish.errors.JellyfishNoElement('Session does not exist')
+		}
+
 		return jellyfish.insertCard(context.session, {
 			type: options.type,
 			links: [],
@@ -90,7 +96,7 @@ module.exports = {
 			data: {
 				timestamp: context.timestamp,
 				target: card.id,
-				actor: context.actor.id,
+				actor: sessionCard.data.actor,
 				payload: options.payload
 			}
 		})
