+++ lib/sdk/kernel.js
@@ -17,37 +17,21 @@
 const _ = require('lodash')
 const debug = require('debug')('jellyfish:kernel')
 const Bluebird = require('bluebird')
-const objectTemplate = require('object-template')
 const EventEmitter = require('events').EventEmitter
 const jsonSchema = require('./json-schema')
 const errors = require('./errors')
 const CARDS = require('./cards')
+const permissionFilter = require('./permission-filter')
 
 const getFilteredQuery = async (kernel, session, schema, options = {}) => {
-	const finalSchema = schema.type === CARDS.view.slug ? kernel.getViewSchema(schema) : schema
-	const modes = options.writeMode ? [ 'read', 'write' ] : [ 'read' ]
-	const user = await kernel.getSessionUser(session)
-	const roles = _.chain([ user.slug ])
-		.concat(user.data.roles)
-		.reduce((accumulator, role) => {
-			Reflect.apply(accumulator.push, accumulator, _.map(modes, (mode) => {
-				return `view-${mode}-${role}`
-			}))
-
-			return accumulator
-		}, [])
-		.value()
-
-	const filters = await kernel.getViewFilters(roles)
-	filters.push(finalSchema)
-
-	return jsonSchema.merge(_.map(filters, (filter) => {
-		return objectTemplate.compile(filter, {
-			user
-		}, {
-			delimiters: [ '\\[', '\\]' ]
-		})
-	}))
+	return permissionFilter.getQuery(kernel.backend, session, schema, {
+		writeMode: options.writeMode,
+		buckets: {
+			session: kernel.getBucketForType('session'),
+			user: kernel.getBucketForType('user'),
+			view: kernel.getBucketForType('view')
+		}
+	})
 }
 
 module.exports = class Kernel {
@@ -127,22 +111,18 @@ module.exports = class Kernel {
 		// Built-in cards
 
 		debug('Upserting minimal required cards')
-		await this.unsafeUpsertCard(CARDS.type)
-		await this.unsafeUpsertCard(CARDS.session)
-		await this.unsafeUpsertCard(CARDS.user)
-		const adminUserId = await this.unsafeUpsertCard(CARDS['user-admin'])
+		const unsafeUpsert = (card) => {
+			return permissionFilter.unsafeUpsertCard(this.backend, this.getBucketForType(card.type), card)
+		}
+
+		await unsafeUpsert(CARDS.type)
+		await unsafeUpsert(CARDS.session)
+		await unsafeUpsert(CARDS.user)
+		const adminUserId = await unsafeUpsert(CARDS['user-admin'])
 
 		this.sessions = {
-			admin: await this.unsafeUpsertCard({
-				slug: 'session-admin-kernel',
-				type: 'session',
-				links: [],
-				tags: [],
-				active: true,
-				data: {
-					actor: adminUserId
-				}
-			})
+			admin: await permissionFilter.createSession(this.backend,
+				this.getBucketForType('session'), adminUserId, 'admin-kernel')
 		}
 
 		return Bluebird.all([
@@ -239,103 +219,6 @@ module.exports = class Kernel {
 	}
 
 	/**
-   * @summary Upsert a card to the kernel in an unsafe way
-   * @function
-   * @private
-	 *
-	 * @description
-	 * This function should only be used to boostrap the system.
-   *
-   * @param {Object} card - card object
-   * @returns {String} the card id
-   *
-   * @example
-   * const kernel = new Kernel(backend, { ... })
-   * await kernel.initialize()
-   *
-   * const id = await kernel.unsafeUpsertCard({ ... })
-   * console.log(id)
-   */
-	async unsafeUpsertCard (card) {
-		jsonSchema.validate(CARDS.card.data.schema, card)
-		jsonSchema.validate(CARDS[card.type].data.schema, card)
-		return this.backend.upsertElement(this.getBucketForType(card.type), card)
-	}
-
-	/**
-	 * @summary Get the user that corresponds to a session
-	 * @function
-	 * @public
-	 *
-	 * @param {String} session - session id
-	 * @returns {Object} user card
-	 *
-	 * @example
-   * const kernel = new Kernel(backend, { ... })
-   * await kernel.initialize()
-   *
-   * const user = await kernel.getSessionUser('4a962ad9-20b5-4dd8-a707-bf819593cc84')
-   * console.log(user.data.email)
-	 */
-	async getSessionUser (session) {
-		// Notice we can't use .query() here since that function depends on this one
-		const sessionCard = await this.backend.getElementById(this.getBucketForType('session'), session)
-		if (!sessionCard) {
-			throw new this.errors.JellyfishNoElement(`Invalid session: ${session}`)
-		}
-
-		if (sessionCard.data.expiration && new Date() > new Date(sessionCard.data.expiration)) {
-			throw new this.errors.JellyfishSessionExpired(`Session expired at: ${sessionCard.data.expiration}`)
-		}
-
-		const actor = await this.backend.getElementById(this.getBucketForType('user'), sessionCard.data.actor)
-		if (!actor) {
-			throw new this.errors.JellyfishNoElement(`Invalid actor: ${sessionCard.data.actor}`)
-		}
-
-		return actor
-	}
-
-	/**
-	 * @summary Get the schema filters that apply to a set of views
-	 * @function
-	 * @private
-	 *
-	 * @param {String[]} roles - roles
-	 * @returns {Object[]} filters
-	 *
-	 * @example
-	 * const kernel = new Kernel(backend, { ... })
-	 * await kernel.initialize()
-	 *
-	 * const filters = await kernel.getViewFilters([ 'view-resineer', 'view-user-guest' ])
-	 *
-	 * for (const filter of filters) {
-	 *   console.log(filter)
-	 * }
-	 */
-	async getViewFilters (roles) {
-		const CARD_TYPE = 'view'
-		return Bluebird.reduce(roles, (accumulator, role) => {
-			return this.backend.getElementBySlug(this.getBucketForType(CARD_TYPE), role)
-				.then((card) => {
-					if (card && card.type !== CARD_TYPE) {
-						return null
-					}
-
-					return this.getViewSchema(card)
-				})
-				.then((schema) => {
-					if (schema) {
-						accumulator.push(schema)
-					}
-
-					return accumulator
-				})
-		}, [])
-	}
-
-	/**
    * @summary Insert a card to the kernel
    * @function
    * @protected
@@ -385,45 +268,6 @@ module.exports = class Kernel {
 		}
 
 		return this.backend.insertElement(bucket, data)
-	}
-
-	/**
-	 * @summary Get the schema of a view card
-	 * @function
-	 * @public
-	 *
-   * @param {String} session - session id
-	 * @param {Object} card - view card
-	 * @returns {(Object|Null} schema
-	 *
-	 * @example
-   * const kernel = new Kernel(backend, { ... })
-   * await kernel.initialize()
-	 *
-	 * const card = await kernel.getCardBySlug('4a962ad9-20b5-4dd8-a707-bf819593cc84', 'view-all')
-	 * const schema = kernel.getViewSchema(card)
-	 * console.log(schema)
-	 */
-	// eslint-disable-next-line class-methods-use-this
-	getViewSchema (card) {
-		if (!card) {
-			return null
-		}
-
-		const conjunctions = _.map(_.get(card, [ 'data', 'allOf' ]), 'schema')
-		const disjunctions = _.map(_.get(card, [ 'data', 'anyOf' ]), 'schema')
-
-		if (_.isEmpty(conjunctions) && _.isEmpty(disjunctions)) {
-			return null
-		}
-
-		if (!_.isEmpty(disjunctions)) {
-			conjunctions.push({
-				anyOf: disjunctions
-			})
-		}
-
-		return jsonSchema.merge(conjunctions)
 	}
 
 	/**
