+++ test/sdk/kernel.spec.js
@@ -65,316 +65,6 @@ ava.test('should be able to disconnect the kernel multiple times without errors'
 	})
 })
 
-ava.test('.getSessionUser() should throw if the session is invalid', async (test) => {
-	await test.throws(test.context.kernel.getSessionUser('xxxxxxxxxxxxxxxxxxxxxxxxxx'), errors.JellyfishNoElement)
-})
-
-ava.test('.getSessionUser() should throw if the session actor is invalid', async (test) => {
-	const session = await test.context.kernel.insertCard(test.context.kernel.sessions.admin, {
-		type: 'session',
-		active: true,
-		links: [],
-		tags: [],
-		data: {
-			actor: '4a962ad9-20b5-4dd8-a707-bf819593cc84'
-		}
-	})
-
-	await test.throws(test.context.kernel.getSessionUser(session), errors.JellyfishNoElement)
-})
-
-ava.test('.getSessionUser() should get the session user given the session did not expire', async (test) => {
-	const id = await test.context.kernel.insertCard(test.context.kernel.sessions.admin, {
-		slug: 'user-johndoe',
-		type: 'user',
-		active: true,
-		links: [],
-		tags: [],
-		data: {
-			email: 'johndoe@example.com',
-			roles: [ 'foo', 'bar' ]
-		}
-	})
-
-	const date = new Date()
-	date.setDate(date.getDate() + 1)
-
-	const session = await test.context.kernel.insertCard(test.context.kernel.sessions.admin, {
-		type: 'session',
-		active: true,
-		links: [],
-		tags: [],
-		data: {
-			actor: id,
-			expiration: date.toISOString()
-		}
-	})
-
-	const user = await test.context.kernel.getSessionUser(session)
-	test.deepEqual(user, Object.assign({
-		id
-	}, user))
-})
-
-ava.test('.getSessionUser() should throw if the session expired', async (test) => {
-	const user = await test.context.kernel.insertCard(test.context.kernel.sessions.admin, {
-		slug: 'user-johndoe',
-		type: 'user',
-		active: true,
-		links: [],
-		tags: [],
-		data: {
-			email: 'johndoe@example.com',
-			roles: [ 'foo', 'bar' ]
-		}
-	})
-
-	const date = new Date()
-	date.setDate(date.getDate() - 1)
-
-	const session = await test.context.kernel.insertCard(test.context.kernel.sessions.admin, {
-		type: 'session',
-		active: true,
-		links: [],
-		tags: [],
-		data: {
-			actor: user,
-			expiration: date.toISOString()
-		}
-	})
-
-	await test.throws(test.context.kernel.getSessionUser(session), errors.JellyfishSessionExpired)
-})
-
-ava.test('.getViewFilters() should return an empty array given no views', async (test) => {
-	const filters = await test.context.kernel.getViewFilters([])
-	test.deepEqual(filters, [])
-})
-
-ava.test('.getViewFilters() should return the schema of a single view', async (test) => {
-	await test.context.kernel.insertCard(test.context.kernel.sessions.admin, {
-		slug: 'view-read-foo',
-		type: 'view',
-		tags: [],
-		links: [],
-		active: true,
-		data: {
-			allOf: [
-				{
-					name: 'Active cards',
-					schema: {
-						type: 'object',
-						properties: {
-							active: {
-								type: 'boolean',
-								const: true
-							}
-						},
-						required: [ 'active' ]
-					}
-				}
-			]
-		}
-	})
-
-	const filters = await test.context.kernel.getViewFilters([ 'view-read-foo' ])
-	test.deepEqual(filters, [
-		{
-			type: 'object',
-			properties: {
-				active: {
-					type: 'boolean',
-					const: true
-				}
-			},
-			required: [ 'active' ]
-		}
-	])
-})
-
-ava.test('.getViewFilters() should ignore undefined views', async (test) => {
-	await test.context.kernel.insertCard(test.context.kernel.sessions.admin, {
-		slug: 'view-read-foo',
-		type: 'view',
-		tags: [],
-		links: [],
-		active: true,
-		data: {
-			allOf: [
-				{
-					name: 'Active cards',
-					schema: {
-						type: 'object',
-						properties: {
-							active: {
-								type: 'boolean',
-								const: true
-							}
-						},
-						required: [ 'active' ]
-					}
-				}
-			]
-		}
-	})
-
-	const filters = await test.context.kernel.getViewFilters([ 'view-hello', 'view-read-foo', 'view-world' ])
-	test.deepEqual(filters, [
-		{
-			type: 'object',
-			properties: {
-				active: {
-					type: 'boolean',
-					const: true
-				}
-			},
-			required: [ 'active' ]
-		}
-	])
-})
-
-ava.test('.getViewFilters() should ignore cards that are not views', async (test) => {
-	await test.context.kernel.insertCard(test.context.kernel.sessions.admin, {
-		slug: 'view-read-hello',
-		type: 'card',
-		tags: [],
-		links: [],
-		active: true,
-		data: {
-			allOf: [
-				{
-					name: 'Action cards',
-					schema: {
-						type: 'object',
-						properties: {
-							type: {
-								type: 'string',
-								const: 'action'
-							}
-						},
-						required: [ 'type' ]
-					}
-				}
-			]
-		}
-	})
-
-	await test.context.kernel.insertCard(test.context.kernel.sessions.admin, {
-		slug: 'view-read-foo',
-		type: 'view',
-		tags: [],
-		links: [],
-		active: true,
-		data: {
-			allOf: [
-				{
-					name: 'Active cards',
-					schema: {
-						type: 'object',
-						properties: {
-							active: {
-								type: 'boolean',
-								const: true
-							}
-						},
-						required: [ 'active' ]
-					}
-				}
-			]
-		}
-	})
-
-	const filters = await test.context.kernel.getViewFilters([ 'view-read-hello', 'view-read-foo' ])
-	test.deepEqual(filters, [
-		{
-			type: 'object',
-			properties: {
-				active: {
-					type: 'boolean',
-					const: true
-				}
-			},
-			required: [ 'active' ]
-		}
-	])
-})
-
-ava.test('.getViewFilters() should return the schemas of two roles', async (test) => {
-	await test.context.kernel.insertCard(test.context.kernel.sessions.admin, {
-		slug: 'view-read-foo',
-		type: 'view',
-		tags: [],
-		links: [],
-		active: true,
-		data: {
-			allOf: [
-				{
-					name: 'Active cards',
-					schema: {
-						type: 'object',
-						properties: {
-							active: {
-								type: 'boolean',
-								const: true
-							}
-						},
-						required: [ 'active' ]
-					}
-				}
-			]
-		}
-	})
-
-	await test.context.kernel.insertCard(test.context.kernel.sessions.admin, {
-		slug: 'view-read-bar',
-		type: 'view',
-		tags: [],
-		links: [],
-		active: true,
-		data: {
-			allOf: [
-				{
-					name: 'Action cards',
-					schema: {
-						type: 'object',
-						properties: {
-							type: {
-								type: 'string',
-								const: 'action'
-							}
-						},
-						required: [ 'type' ]
-					}
-				}
-			]
-		}
-	})
-
-	const filters = await test.context.kernel.getViewFilters([ 'view-read-foo', 'view-read-bar' ])
-	test.deepEqual(filters, [
-		{
-			type: 'object',
-			properties: {
-				active: {
-					type: 'boolean',
-					const: true
-				}
-			},
-			required: [ 'active' ]
-		},
-		{
-			type: 'object',
-			properties: {
-				type: {
-					type: 'string',
-					const: 'action'
-				}
-			},
-			required: [ 'type' ]
-		}
-	])
-})
-
 ava.test('.insertCard() should throw an error if the element is not a valid card', async (test) => {
 	await test.throws(test.context.kernel.insertCard(test.context.kernel.sessions.admin, {
 		hello: 'world'
@@ -796,430 +486,6 @@ ava.test('.getCardById() should return an inactive card by its id', async (test)
 		tags: [],
 		data: {}
 	})
-})
-
-ava.test('.getViewSchema() should return null if the card is not a view', (test) => {
-	const schema = test.context.kernel.getViewSchema(CARDS['user-admin'])
-	test.deepEqual(schema, null)
-})
-
-ava.test('.getViewSchema() should preserve template interpolations in user properties', (test) => {
-	const schema = test.context.kernel.getViewSchema({
-		type: 'view',
-		links: [],
-		tags: [],
-		active: true,
-		data: {
-			allOf: [
-				{
-					name: 'foo',
-					schema: {
-						type: 'object',
-						properties: {
-							foo: {
-								type: 'string',
-								const: '[user.slug]'
-							}
-						},
-						required: [ 'foo' ]
-					}
-				}
-			]
-		}
-	})
-
-	test.deepEqual(schema, {
-		type: 'object',
-		properties: {
-			foo: {
-				type: 'string',
-				const: '[user.slug]'
-			}
-		},
-		required: [ 'foo' ]
-	})
-})
-
-ava.test('.getViewSchema() should preserve template interpolations in schema properties', (test) => {
-	const schema = test.context.kernel.getViewSchema({
-		type: 'view',
-		links: [],
-		tags: [],
-		active: true,
-		data: {
-			allOf: [
-				{
-					name: 'foo',
-					schema: {
-						type: 'object',
-						properties: {
-							foo: {
-								type: '[user.type]'
-							}
-						},
-						required: [ 'foo' ]
-					}
-				}
-			]
-		}
-	})
-
-	test.deepEqual(schema, {
-		type: 'object',
-		properties: {
-			foo: {
-				type: '[user.type]'
-			}
-		},
-		required: [ 'foo' ]
-	})
-})
-
-ava.test('.getViewSchema() should return a schema given a view card with two conjunctions', (test) => {
-	const schema = test.context.kernel.getViewSchema({
-		type: 'view',
-		links: [],
-		tags: [],
-		active: true,
-		data: {
-			allOf: [
-				{
-					name: 'foo',
-					schema: {
-						type: 'object',
-						properties: {
-							foo: {
-								type: 'string',
-								minLength: 1
-							}
-						},
-						required: [ 'foo' ]
-					}
-				},
-				{
-					name: 'bar',
-					schema: {
-						type: 'object',
-						properties: {
-							foo: {
-								type: 'string',
-								maxLength: 5
-							}
-						},
-						required: [ 'foo' ]
-					}
-				}
-			]
-		}
-	})
-
-	test.deepEqual(schema, {
-		type: 'object',
-		properties: {
-			foo: {
-				type: 'string',
-				minLength: 1,
-				maxLength: 5
-			}
-		},
-		required: [ 'foo' ]
-	})
-})
-
-ava.test('.getViewSchema() should return a schema given a view card with two conjunctions and empty disjunctions', (test) => {
-	const schema = test.context.kernel.getViewSchema({
-		type: 'view',
-		links: [],
-		tags: [],
-		active: true,
-		data: {
-			anyOf: [],
-			allOf: [
-				{
-					name: 'foo',
-					schema: {
-						type: 'object',
-						properties: {
-							foo: {
-								type: 'string',
-								minLength: 1
-							}
-						},
-						required: [ 'foo' ]
-					}
-				},
-				{
-					name: 'bar',
-					schema: {
-						type: 'object',
-						properties: {
-							foo: {
-								type: 'string',
-								maxLength: 5
-							}
-						},
-						required: [ 'foo' ]
-					}
-				}
-			]
-		}
-	})
-
-	test.deepEqual(schema, {
-		type: 'object',
-		properties: {
-			foo: {
-				type: 'string',
-				minLength: 1,
-				maxLength: 5
-			}
-		},
-		required: [ 'foo' ]
-	})
-})
-
-ava.test('.getViewSchema() should return a schema given a view card with two disjunctions', (test) => {
-	const schema = test.context.kernel.getViewSchema({
-		type: 'view',
-		links: [],
-		tags: [],
-		active: true,
-		data: {
-			anyOf: [
-				{
-					name: 'foo',
-					schema: {
-						type: 'object',
-						properties: {
-							type: {
-								type: 'string',
-								const: 'view'
-							}
-						},
-						required: [ 'type' ]
-					}
-				},
-				{
-					name: 'bar',
-					schema: {
-						type: 'object',
-						properties: {
-							type: {
-								type: 'string',
-								const: 'action'
-							}
-						},
-						required: [ 'type' ]
-					}
-				}
-			]
-		}
-	})
-
-	test.deepEqual(schema, {
-		type: 'object',
-		anyOf: [
-			{
-				type: 'object',
-				properties: {
-					type: {
-						type: 'string',
-						const: 'view'
-					}
-				},
-				required: [ 'type' ]
-			},
-			{
-				type: 'object',
-				properties: {
-					type: {
-						type: 'string',
-						const: 'action'
-					}
-				},
-				required: [ 'type' ]
-			}
-		]
-	})
-})
-
-ava.test('.getViewSchema() should return a schema given a view card with two disjunctions and empty conjunctions', (test) => {
-	const schema = test.context.kernel.getViewSchema({
-		type: 'view',
-		links: [],
-		tags: [],
-		active: true,
-		data: {
-			allOf: [],
-			anyOf: [
-				{
-					name: 'foo',
-					schema: {
-						type: 'object',
-						properties: {
-							type: {
-								type: 'string',
-								const: 'view'
-							}
-						},
-						required: [ 'type' ]
-					}
-				},
-				{
-					name: 'bar',
-					schema: {
-						type: 'object',
-						properties: {
-							type: {
-								type: 'string',
-								const: 'action'
-							}
-						},
-						required: [ 'type' ]
-					}
-				}
-			]
-		}
-	})
-
-	test.deepEqual(schema, {
-		type: 'object',
-		anyOf: [
-			{
-				type: 'object',
-				properties: {
-					type: {
-						type: 'string',
-						const: 'view'
-					}
-				},
-				required: [ 'type' ]
-			},
-			{
-				type: 'object',
-				properties: {
-					type: {
-						type: 'string',
-						const: 'action'
-					}
-				},
-				required: [ 'type' ]
-			}
-		]
-	})
-})
-
-ava.test('.getViewSchema() should return a schema given a view card with two disjunctions and two conjunctions', (test) => {
-	const schema = test.context.kernel.getViewSchema({
-		type: 'view',
-		links: [],
-		tags: [],
-		active: true,
-		data: {
-			anyOf: [
-				{
-					name: 'foo',
-					schema: {
-						type: 'object',
-						properties: {
-							type: {
-								type: 'string',
-								const: 'view'
-							}
-						},
-						required: [ 'type' ]
-					}
-				},
-				{
-					name: 'bar',
-					schema: {
-						type: 'object',
-						properties: {
-							type: {
-								type: 'string',
-								const: 'action'
-							}
-						},
-						required: [ 'type' ]
-					}
-				}
-			],
-			allOf: [
-				{
-					name: 'foo',
-					schema: {
-						type: 'object',
-						properties: {
-							foo: {
-								type: 'string',
-								minLength: 1
-							}
-						},
-						required: [ 'foo' ]
-					}
-				},
-				{
-					name: 'bar',
-					schema: {
-						type: 'object',
-						properties: {
-							foo: {
-								type: 'string',
-								maxLength: 5
-							}
-						},
-						required: [ 'foo' ]
-					}
-				}
-			]
-		}
-	})
-
-	test.deepEqual(schema, {
-		type: 'object',
-		properties: {
-			foo: {
-				type: 'string',
-				minLength: 1,
-				maxLength: 5
-			}
-		},
-		required: [ 'foo' ],
-		anyOf: [
-			{
-				type: 'object',
-				properties: {
-					type: {
-						type: 'string',
-						const: 'view'
-					}
-				},
-				required: [ 'type' ]
-			},
-			{
-				type: 'object',
-				properties: {
-					type: {
-						type: 'string',
-						const: 'action'
-					}
-				},
-				required: [ 'type' ]
-			}
-		]
-	})
-})
-
-ava.test('.getViewSchema() should return null given a view card with no filters', (test) => {
-	const schema = test.context.kernel.getViewSchema({
-		type: 'view',
-		links: [],
-		tags: [],
-		active: true,
-		data: {}
-	})
-
-	test.deepEqual(schema, null)
 })
 
 ava.test('.query() should return the cards that match a schema', async (test) => {
